rename chart options to chart menu
